update: use new conan generators + remove MSVC warnings

+++ src/game/Formulas.cpp
@@ -9,5 +9,5 @@
 
 uint8_t Formulas::getLevelFromXp(uint64_t xp)
 {
-    return 2 + std::log2l(static_cast<long double>(xp) / 10'000);
+    return static_cast<uint8_t>(2 + std::log2l(static_cast<long double>(xp) / 10'000));
 }
